% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_helpers.R
\name{render_list_of_authors}
\alias{render_list_of_authors}
\title{Generate list of authors}
\usage{
render_list_of_authors(input, output = NULL)
}
\arguments{
\item{input}{Input tsv file <filename.tsv>}

\item{output}{Output docx file. If not specified, <filename.docx> is used.}
}
\description{
Generate MS Word doc with formatted author list and formatted affiliations
from input tsv file

Assumptions:
- input file contains columns 'Author name' and 'affiliation 1', 'affiliation 2', ...
- rows are sorted by desired occurence of authors on manuscript
- affiliation indices per author are sorted numerically
}
