% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datascience_helpers.R
\name{setup_project}
\alias{setup_project}
\title{Setup internals of data science project}
\usage{
setup_project()
}
\description{
This function is meant to run within a new R project directory to
1. Consistently set up the directory structure
2. Initialize new project-local environment with renv
}
